domain driven design refactoring
