+++ apps/api/src/index.ts
@@ -2,28 +2,24 @@ import { config as loadDotFile } from "dotenv";
 loadDotFile();
 
 import express from "express";
-import rateLimit from "express-rate-limit";
 import connectToDatabase from "./config/db";
 import passport from "passport";
 import mediaRoutes from "./media/routes";
 import signatureRoutes from "./signature/routes";
 import mediaSettingsRoutes from "./media-settings/routes";
 import tusRoutes from "./tus/routes";
+import mcpRoutes from "./mcp/routes";
 import logger from "./services/log";
 import { createUser, findByEmail } from "./user/queries";
 import { Apikey, User } from "@medialit/models";
 import { getApiKeyByUserId } from "./apikey/queries";
 import swaggerUi from "swagger-ui-express";
 import swaggerOutput from "./swagger_output.json";
-import { mcpAuth } from "./auth/middleware";
-import { oauthRouter } from "./oauth/server";
 
 import { spawn } from "child_process";
 import { cleanupTUSUploads } from "./tus/cleanup";
 import { cleanupExpiredTempUploads } from "./media/cleanup";
 import { HOUR_IN_SECONDS } from "./config/constants";
-import { createMCPSession } from "./mcp/server";
-import { StreamableHTTPServerTransport } from "@modelcontextprotocol/sdk/server/streamableHttp.js";
 
 connectToDatabase();
 const app = express();
@@ -94,6 +90,7 @@ app.use("/settings/media", mediaSettingsRoutes(passport));
 app.use("/media/signature", signatureRoutes);
 app.use("/media", tusRoutes);
 app.use("/media", mediaRoutes);
+app.use(mcpRoutes);
 
 app.get(
     "/cleanup/temp",
@@ -115,109 +112,6 @@ app.get(
         });
     },
 );
-
-const mcpLimiter = rateLimit({
-    windowMs: 60_000,
-    max: 60,
-    standardHeaders: true,
-    legacyHeaders: false,
-    message: {
-        error: "too_many_requests",
-        error_description: "Too many requests.",
-    },
-});
-
-// Active MCP sessions: sessionId → transport
-const mcpSessions = new Map<string, StreamableHTTPServerTransport>();
-
-// CORS middleware for MCP and OAuth endpoints
-const mcpCors = (req: any, res: any, next: any) => {
-    const origin = req.headers.origin || "*";
-    res.header("Access-Control-Allow-Origin", origin);
-    res.header("Access-Control-Allow-Methods", "GET, POST, DELETE, OPTIONS");
-    res.header(
-        "Access-Control-Allow-Headers",
-        "Content-Type, Accept, Mcp-Session-Id, Mcp-Protocol-Version, x-medialit-apikey, Authorization",
-    );
-    res.header("Access-Control-Expose-Headers", "Mcp-Session-Id");
-    if (req.method === "OPTIONS") {
-        return res.status(204).end();
-    }
-    next();
-};
-
-// OAuth endpoints + CORS (no auth required — these are the auth flow itself)
-app.use(["/.well-known", "/oauth"], mcpCors);
-app.use(oauthRouter);
-
-app.post("/mcp", mcpCors, mcpLimiter, mcpAuth, async (req: any, res: any) => {
-    // The MCP SDK (via @hono/node-server) reads rawHeaders to build the Web Standard
-    // Request, so we must patch both req.headers AND req.rawHeaders.
-    // The SDK requires Accept to include BOTH application/json and text/event-stream.
-    const accept = req.headers.accept || "";
-    const needsJson = !accept.includes("application/json");
-    const needsSSE = !accept.includes("text/event-stream");
-    if (needsJson || needsSSE) {
-        const additions: string[] = [];
-        if (needsJson) additions.push("application/json");
-        if (needsSSE) additions.push("text/event-stream");
-        const newAccept = accept
-            ? `${accept}, ${additions.join(", ")}`
-            : additions.join(", ");
-        req.headers.accept = newAccept;
-        // Patch rawHeaders (used by @hono/node-server to build the Web Standard Request)
-        const rawHeaders: string[] = req.rawHeaders;
-        let found = false;
-        for (let i = 0; i < rawHeaders.length; i += 2) {
-            if (rawHeaders[i].toLowerCase() === "accept") {
-                rawHeaders[i + 1] = newAccept;
-                found = true;
-                break;
-            }
-        }
-        if (!found) rawHeaders.push("Accept", newAccept);
-    }
-
-    const auth = {
-        token: req.apikey || "",
-        clientId: String(req.userId || req.user?._id || req.user?.id || ""),
-        user: req.user,
-        scopes: [] as string[],
-    };
-
-    const sessionId = req.headers["mcp-session-id"] as string | undefined;
-
-    if (sessionId) {
-        // Route to an existing session
-        const transport = mcpSessions.get(sessionId);
-        if (!transport) {
-            return res.status(404).json({
-                jsonrpc: "2.0",
-                error: { code: -32001, message: "Session not found" },
-                id: null,
-            });
-        }
-        await transport.handleRequest(
-            Object.assign(req, { auth }),
-            res,
-            req.body,
-        );
-    } else {
-        // New session — create a dedicated transport + server pair
-        const transport = createMCPSession(
-            (id) => mcpSessions.set(id, transport),
-            (id) => mcpSessions.delete(id),
-        );
-        await transport.handleRequest(
-            Object.assign(req, { auth }),
-            res,
-            req.body,
-        );
-    }
-});
-
-// CORS preflight for ChatGPT/web-based MCP clients
-app.options("/mcp", mcpCors);
 
 const port = process.env.PORT || 80;
